Revert "fix: rename filename (#4)"
This reverts commit b816e8e10811cb600bf26c84b4517202a5a29fad.
