--- README.md
@@ -1,11 +1,0 @@
-# fixdevcontainer
-
-Sort top-level keys in devcontainer.json
-
-Inspired by [fixpack](https://www.npmjs.com/package/fixpack)!
-
-## Usage
-
-```shell
-npx fixdevcontainer
-```
